Tests/LibWeb/Text/expected/WebAudio/AudioWorkletNode-construct.txt

port is MessagePort: true
port.postMessage: function
parameters.size: 2
parameters.get: function
PASS
